fix: move mimeType to @isomorphic
