+++ packages/trace-viewer/src/ui/attachmentsTab.tsx
@@ -21,12 +21,7 @@ import type { MultiTraceModel } from './modelUtil';
 import { PlaceholderPanel } from './placeholderPanel';
 import type { AfterActionTraceEventAttachment } from '@trace/trace';
 
-// TODO: remove local function and restore import
-// import { isTextualMimeType } from 'playwright-core/lib/utils';
-
-function isTextualMimeType(mimeType: string) {
-  return !!mimeType.match(/^(text\/.*?|application\/(json|(x-)?javascript|xml.*?|ecmascript|graphql|x-www-form-urlencoded)|image\/svg(\+xml)?|application\/.*?(\+json|\+xml))(;\s*charset=.*)?$/);
-}
+import { isTextualMimeType } from '@isomorphic/mimeType';
 
 
 type Attachment = AfterActionTraceEventAttachment & { traceUrl: string };
